pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "your_package"
version = "0.1.0"
description = "An example package"
authors = [
  {name = "Pradyun Gedam", email = "pradyun@example.com"},
]
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = ["mysql-connector-python==8.0.29"]


[project.packages]
include = ["main.py"]

[project.data-files]
data = ["data/*.dat"]
config = ["config/*.conf"]
images = ["images/*.png"]
